requirements.txt
discord.py
requests
python-dotenv
asyncio
openai
beautifulsoup4
lxml
PyPDF2
pycryptodome

# sudo service docker start
# docker build -t asia-northeast1-docker.pkg.dev/rtree-377503/cloudrun-mochi-bot-discord/mochi-bot-discord .
# DISPLAY=":0" gcloud auth login
# gcloud auth configure-docker asia-northeast1-docker.pkg.dev
# docker tag mochi-bot-discord asia-northeast1-docker.pkg.dev/rtree-377503/cloudrun-mochi-bot-discord/mochi-bot-discord
# docker push asia-northeast1-docker.pkg.dev/rtree-377503/cloudrun-mochi-bot-discord/mochi-bot-discord
# gcloud auth revoke
#
#
#
